chore(test): batch-scheduler.spec.ts 분할 — 2,251줄 단일 spec 구조 개선
93개 테스트를 11개 spec 파일로 분리:
- batch-scheduler.test-setup.ts: 공용 헬퍼 (executionCoordinator, DEFAULT_SCHEDULER_CONFIG)
- env-defaults / constructor-config / lifecycle / status-diagnostics
- run-job / job-execution-retry / concurrency-queue
- maintenance-jobs / error-handling / logging / relation-validation

각 파일 500줄 이하, 테스트 의미·통과 조건 동일. Related #166
